+++ 03-frontend/angular-ecommerce/src/app/common/product.ts
@@ -2,6 +2,7 @@
 // will be read in from remote API
 export class Product {
 
+  id: string;
   sku: string;
   name: string;
   description: string;