+++ 03-frontend/angular-ecommerce/src/app/services/product.service.ts
@@ -18,12 +18,24 @@ export class ProductService {
   // inject httpClient into ProductService
   constructor(private httpClient: HttpClient) { }
 
+  // gets a list of products
   getProductList(theCategoryId: number): Observable<Product[]> {
 
     // need to build URL based on category id 
     const searchUrl = `${this.baseUrl}/search/findByCategoryId?id=${theCategoryId}`;
 
     return this.getProducts(searchUrl);
+  }
+
+  // for pagination
+  getProductListPaginate(thePage: number, thePageSize: number, 
+                         theCategoryId: number): Observable<GetResponseProducts> {
+
+    // need to build URL based on category id, page and size
+    const searchUrl = `${this.baseUrl}/search/findByCategoryId?id=${theCategoryId}`
+                     + `&page=${thePage}&size=${thePageSize}`;
+
+    return this.httpClient.get<GetResponseProducts>(searchUrl);
   }
 
   searchProducts(theKeyword: string): Observable<Product[]> {
@@ -70,6 +82,12 @@ export class ProductService {
 interface GetResponseProducts {
   _embedded: {
     products: Product[];
+  },
+  page: {
+    size: number,
+    totalElements: number,
+    totalPages: number,
+    number: number
   }
 }